replace int with long

+++ flight test/test1_1/test1_1.ino
@@ -1,5 +1,5 @@
 #define RC_pin 18 //PPM
-#define button_pin 14//A0
+#define button_pin 14 //A0
 
 void setup()
 {
@@ -10,7 +10,7 @@ void setup()
 void loop()
 {
     int i; //ループの数
-    int time1,time2;
+    long time1,time2;
     //int PPMMODE_example[8] = {A,E,T,R,C,R,A,?}
     //エレベーター上げるが1900でエレベーター下げるが1100
     //ラダーは知らん
